patch(mutators): move `GenerativeMutator` class out of `core.py`
